requirements.txt
python-telegram-bot==20.7
sqlalchemy
python-dotenv
psycopg2-binary
openai
pandas